AISEC-DRS and ICLR-DRS renamed to Fixed Size Chunk DRS and K-Partition DRS, respectively
