+++ src/maltorch/data_processing/dynamic_sequential_drs_preprocessing.py
@@ -3,7 +3,7 @@ from secmlt.models.data_processing.data_processing import DataProcessing
 import math
 
 
-class SequentialDeRandomizedPreprocessing(DataProcessing):
+class DynamicSequentialDeRandomizedPreprocessing(DataProcessing):
     """
     Daniel Gibert, Giulio Zizzo, Quan Le, Jordi Planes
     Adversarial Robustness of Deep Learning-based Malware Detectors via (De) Randomized Smoothing
