+++ src/maltorch/datasets/dynamic_random_drs_dataset.py
@@ -1,15 +1,14 @@
 from typing import Tuple
 
 import torch
+from random import randint
 from maltorch.datasets.dynamic_drs_dataset import DynamicChunkSizeDRSDataset
-import math
 
-
-class SequentialDRSDataset(DynamicChunkSizeDRSDataset):
+class DynamicRandomDRSDataset(DynamicChunkSizeDRSDataset):
     """
     Daniel Gibert, Giulio Zizzo, Quan Le, Jordi Planes
     Adversarial Robustness of Deep Learning-based Malware Detectors via (De) Randomized Smoothing
-    IEEE Access 2024 -Sequential Chunks-based (De)Randomized Smoothing
+    IEEE Access 2024 - Random Chunks-based (De)Randomized Smoothing
     """
     def __init__(self,
                  csv_filepath: str = None,
@@ -37,33 +36,27 @@ class SequentialDRSDataset(DynamicChunkSizeDRSDataset):
             min_chunk_size=min_chunk_size
         )
 
-    def generate_testing_examples(self, batch) -> Tuple[torch.Tensor, torch.Tensor]:
+    def generate_testing_examples(self, batch: list) -> Tuple[torch.Tensor, torch.Tensor]:
         """
         During testing, given an input example we sequentially extract N chunks of bytes. This chunks may overlap with
         one another depending on the amount of chunks you want to extract and the size of each chunk
         """
-
         if len(batch) == 1:  # Only implemented for batch sizes equals to 1
+            # Code here
             x = batch[0][0]
             max_size = x.size()[0]
-
-            group_size = math.ceil(max_size * self.file_percentage)
-            group_size = max(self.min_chunk_size, group_size) # The chunk size has to be at least equal to the kernel size of the first convolutional layer
-            overlap_size = group_size - int(max_size / self.num_chunks)
-
-            to_substract_A = math.ceil((group_size - overlap_size) / self.num_chunks)
-            to_substract_B = math.ceil((group_size - overlap_size) * self.file_percentage)
-            to_substract = to_substract_B - to_substract_A
+            chunk_size = int(-(-(max_size * self.file_percentage) // 1))  # Used to round up the float value
+            chunk_size = max(self.min_chunk_size, chunk_size) # The chunk size has to be at least equal to the kernel size of the first convolutional layer
 
             vecs = []
-            for i in range(self.num_chunks):
-                start = i * (group_size - overlap_size - to_substract)
-                end = start + group_size
-                vecs.append(x[start:end])
+            for _ in range(self.num_chunks):
+                curr_idx = randint(0, len(x) - chunk_size)
+                vecs.append(x[curr_idx:curr_idx + chunk_size])
 
             x = torch.nn.utils.rnn.pad_sequence(vecs, batch_first=True, padding_value=self.padding_idx)
             y = batch[0][1]
             return x, y
         else:
             raise NotImplementedError
+
